refactor: splited validateCompany into two other usecases
